backend/requirements.txt
django>=4.2,<5.0
djangorestframework
psycopg2-binary
django-cors-headers
google-generativeai
python-dotenv
dj-database-url
